server/print_really/pyproject.toml
[tool.pdm]

[project]
name = "for_printing"
version = "0.1.0"
description = ""
authors = [
    {name = "dayi", email = "yumibiao10@gmail.com"},
]
dependencies = [
    "flask>=2.2.3",
    "pillow>=9.5.0",
    "docx2pdf>=0.1.8",
    "requests>=2.28.2",
    "aspose-words>=23.4.0",
    "PyYAML>=6.0",
    "aiohttp>=3.8.4",
    "asyncio>=3.4.3",
]
requires-python = ">=3.7,<=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
